standard import pointer matchCount
