+++ models/bafflerz.go
@@ -1,20 +1,15 @@
 package models
 
-//Baffler Define universal methods for a Baffler
-type Baffler interface {
-	ProcessFile(file string, rules []UpdateRule) []string
+// Baffler A structure to mystify a file
+type Baffler struct {
+	Name      string
+	Rules     []RuleSet
+	Content   *[]byte
+	Injection map[string]string
 }
 
-//UpdateRule Each change to a file has a condition and function to perform that update
-//For example condition: import is in line, then update the import statements by aliasing them
-type UpdateRule interface {
-	Condition(s *string) bool
-	Update(s *string) string
-}
-
-//BaseBaffler All bafflers, regardless of language or obfuscation method will have
-type BaseBaffler struct {
-	Injection   map[string]string
-	ChangeCount uint32
-	rules       []UpdateRule
+// RuleSet A pair of functions to identify rules and update the text with the rule
+type RuleSet interface {
+	Find(c *[]byte)
+	Update(c *[]byte)
 }
